#11 put tileuploader inside of tileset editor
